Forward cookbook_name to remote_file resource
The JenkinsWindowsSlave provider is not using the DSL to instanciate the
'remote_file' resource. This means some creation steps are bypassed and
some attributes are not set (like 'cookbook_name' which is useful for us).

Other attributes will still be missing, I wanted this change to be as
small as possible but I can use the DSL if needed.